celery_task_app/requirements.txt
celery[redis]
tensorflow==2.14.0
tensorrt
pillow
numpy